refactor - 패키지 구조 조정
- manage 하위에 있던 패키지를 main 패키지 밑으로 이동
