+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/respond/RespondPlanContract.kt
@@ -1,4 +1,4 @@
-package com.yapp.growth.presentation.ui.main.manage.respond
+package com.yapp.growth.presentation.ui.main.respond
 
 import com.yapp.growth.base.ViewEvent
 import com.yapp.growth.base.ViewSideEffect
